:truck: move PricingProvider on to global context
